separate the 2D detailed views of tracker panels to a new module
